+++ src/features/common/components/Card/styles.css.ts
