+++ src/features/home/components/CardList/CardList.tsx
@@ -2,7 +2,7 @@ import { Flex } from '@/utils/Flex'
 import { type Post } from '@/features/post/types/post'
 import { Txt } from '@/common/components/Typography'
 import * as style from './style.css'
-import { Card } from '@/common/Card'
+import { Card } from '@/common/components/Card'
 
 export const CardList = ({ posts }: { posts: Post[] }) => {
   return (
